requirements.txt
# AI Trading System - Production Requirements (Pinned)
# ===================================================
# Rationale: For deterministic builds & reproducibility across environments.
# Versions aligned with service Dockerfiles (Python 3.11). When upgrading:
# 1. Update a staging branch, 2. Rebuild all service images, 3. Run integration tests.
# NOTE: psycopg3 vs psycopg2 inconsistency: containers currently use psycopg2-binary.
#       TODO(persistence): Migrate fully to psycopg (v3) or standardize on psycopg2.

# ---- Core Web Framework ----
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.9.0
pydantic-settings==2.5.2
jinja2==3.1.4  # Server-side rendered dashboards

# ---- Database and Caching ----
redis[hiredis]==5.2.1
psycopg[binary,pool]==3.1.19  # Not used in current images (see note above)
psycopg2-binary==2.9.9
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1

# ---- HTTP and WebSocket ----
httpx==0.27.2
# websockets pin must remain <11 due to alpaca-trade-api (>=9,<11)
websockets==10.4
aiohttp==3.11.10

# ---- Authentication and Security ----
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1  # Pin to 4.x for passlib 1.7.4 compatibility
python-multipart==0.0.6
cryptography==42.0.7
PyJWT==2.8.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.4.0

# ---- Monitoring and Logging ----
structlog==24.1.0
prometheus-client==0.21.0
# Align OpenTelemetry versions with exporter-prometheus (0.47b0 expects ~1.26.x)
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-prometheus==0.47b0

# ---- Data Processing and ML ----
numpy==1.26.4
pandas==2.2.3
polars==0.20.31
pyarrow==15.0.2
scipy==1.14.1
scikit-learn==1.5.2
xgboost==2.0.3
lightgbm==4.3.0

# ---- Deep Learning ----
torch==2.1.0
torch-geometric==2.4.0
transformers==4.35.0
sentence-transformers==2.2.2

# ---- Trading APIs ----
alpaca-trade-api==3.2.0
polygon-api-client==1.14.2
yfinance==0.2.51
ccxt==4.4.24
ta==0.11.0

# ---- Advanced Analytics ----
networkx==3.2.1
statsmodels==0.14.1
arch==6.3.0
hurst==0.0.5
pywavelets==1.5.0
hmmlearn==0.3.2

# ---- Message Queue ----
# Avro schemas required by our Pulsar messaging layer
pulsar-client[avro]==3.5.0
questdb==2.0.3
aiofiles==23.2.1
matplotlib==3.9.2


# ---- Vector Database ----
weaviate-client==3.26.2

# ---- News and Social APIs (optional) ----
newsapi-python==0.2.7
tweepy==4.14.0
praw==7.7.1

# ---- Secrets Management ----
hvac==2.3.0

# ---- Development and Testing ----
pytest==7.4.4
pytest-asyncio==0.23.5
pytest-cov==4.1.0
black==23.12.1
mypy==1.9.0
pylint==3.1.0
pre-commit==3.6.2

# ---- Utilities ----
python-dateutil==2.8.2
pytz==2024.1
pyyaml==6.0.1
python-dotenv==1.0.1
click==8.1.7
psutil==5.9.8  # System monitoring for god-mode dashboard
docker==7.0.0  # Container management for admin controls

# End of file